% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanText.R
\name{cleanText}
\alias{cleanText}
\title{cleanText Function}
\usage{
cleanText(text)
}
\value{
returns a list of characters
}
\description{
This function cleans the text data. When the string of texts (df$text) is run through this function, the function will remove text after the "@" symbol, "https://", and "http://". Then it removes puncuation, numbers, tabs, back space at the beginning, and graphical texts. Then it cleans up sentences by removing ASCII-control characters and digits.
}
\examples{
cleanText(df$text)

}
\keyword{cleanText}
