% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_df.R
\name{freq_df}
\alias{freq_df}
\title{freq_df Function}
\usage{
freq_df(text)
}
\value{
returns a dataframe of characters
}
\description{
This function transform the text data in a corpus. Then it cleans the corpus by turning the text into lowercase, removing punction, number, and "SMART" stopwords. It then returns a simple frequency dataframe.
}
\examples{
freq_df(df$text)

}
\keyword{cleanCorpus}
