% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_RF.R
\name{mice_RF}
\alias{mice_RF}
\title{mice_RF Function}
\usage{
mice_RF(x)
}
\value{
returns a data frame
}
\description{
This function fills your data that contain NA values with predicted values using the random forest method. m argument is automatically set as 5 meaning it will cycle through the imputing prediction process for 5 times.
}
\examples{
mice_RF(df)

}
\keyword{mice_RF}
