% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindMe.R
\name{remindMe}
\alias{remindMe}
\title{remindMe Function}
\usage{
remindMe()
}
\value{
returns a character string
}
\description{
Sick and tired of looking up how to format date and time? Well, I am. This function tells you which code means what inside of as.Date() function or format() function etc.
}
\examples{
remindMe()

}
\keyword{remindMe}
