% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_mean.R
\name{replace_mean}
\alias{replace_mean}
\title{replace_mean Function}
\usage{
replace_mean(x)
}
\value{
returns a data frame
}
\description{
This function fills your data that contain NA values with each column's mean values.
}
\examples{
replace_mean(df)

}
\keyword{replace_mean}
