% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_median.R
\name{replace_median}
\alias{replace_median}
\title{replace_median Function}
\usage{
replace_median(x)
}
\value{
returns a data frame
}
\description{
This function fills your data that contain NA values with each column's mean values.
}
\examples{
replace_median(df)

}
\keyword{replace_median}
