% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_median.R
\name{return_median}
\alias{return_median}
\title{replace_median Function}
\usage{
return_median(x)
}
\value{
returns string out put
}
\description{
This function returns median and mean values of a dataset.
}
\examples{
return_median(airquality)
> return_median(air)
1 median:  31.5   mean:  39.55882 
2 median:  205   mean:  186.8039 
3 median:  9.7   mean:  9.957516 
4 median:  79   mean:  77.88235 
5 median:  7   mean:  6.993464 
6 median:  16   mean:  15.80392 

}
\keyword{return_median}
